Add Insights/Audits toggle banner to report
Adds a more prominent banner that can be clicked to toggle the view. We
do not store this setting locally at all, so by default any future
report will show audits by default.

Feedback welcome in particular as this is my first patch into LH. I know
that all the other SVG icons are done as background-images but in this
case I needed to control the fill color, hence why I inlined it.

The exact text is still TBC, but I wanted to get this up now for initial
review.